+++ apps/chaingraph-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @badaitech/chaingraph-backend
 
+## 0.1.18
+
+### Patch Changes
+
+- Improve the sidebar execution events card to be more informative and easy to understand. Improve port doc tooltips design and colors, fix several bugs.
+- Updated dependencies
+  - @badaitech/chaingraph-nodes@0.1.18
+  - @badaitech/chaingraph-trpc@0.1.18
+  - @badaitech/chaingraph-types@0.1.18
+
 ## 0.1.17
 
 ### Patch Changes